Fix bug in NekosClient.cs
As of right now entering an invalid URL in https://api.nekos.dev/api/v3/images/ returns a JSON message with a list of possible entries that the client should be using, and a status code of 404 inside of the JSON.

However, this causes `HttpResponseMessage.EnsureSuccessStatusCode()` used in `line 66` of `NekosClient.cs` to not throw an exception since the server technically sent back something with a *HTTP* 200 code, even though the returned *JSON* states that this is a 404 resource not found error. 

This, in turn, causes `NekoClient.SendRequest()` to return a Request instance that apparently indicates success... except that the request's ImageUrl property is null, when it shouldn't even return a successful Request instance at all.

The fix I just did is to simply check the property `status.code` (which always seems to be returned in any request that doesn't bring up a 404 HTTP error) present in the JSON returned by the server *as it does not seem to be checked against inside the code* and to throw an exception when the code is not 200.